requirements.txt
Django
psycopg[binary]
sentry-sdk[django]
Pillow
django-filter
djangorestframework
drf-writable-nested
gunicorn
psycopg2-binary
dj-database-url
django-adminplus
boto3
django-storages
django-storages[s3]
django-admin-rangefilter
